Added Battle Music
I've added the battle music to the music folder. If needed It can be cut in half as the recording has the music repeat a single time. Also, if needed I can add an extra section to the music so it doesn't get stale.
Also renamed the music file names.